Relocate 8 verifier intent/policy helpers to verifier_orchestration
Phase 2 follow-up PR (batch 2) for Issue #682. Moves 8 small/medium
pure-fn verifier helpers (~90 LOC of bodies) out of turn.rs into
verifier_orchestration.rs.

Moved (turn.rs -> verifier_orchestration.rs):

  Production helpers:
    - verifier_framework_signal_for_context (8 LOC)
    - verifier_setup_policy_message (8 LOC)
    - verifier_repair_intent_limits (7 LOC)

  Test-only helpers (#[cfg(test)] preserved):
    - verifier_repair_invalid_can_continue (18 LOC)
    - verifier_repair_decision_targets_hint (13 LOC)
    - repair_intent_edit_payloads (12 LOC)
    - verifier_repair_intent_fingerprint (7 LOC)
    - verifier_repair_intents_fingerprint (11 LOC)

All 8 are promoted to `pub(super)` (file-private before). The 5
test-only helpers retain `#[cfg(test)]` since they were test-only in
turn.rs. turn.rs splits the verifier_orchestration imports into a
production block and a `#[cfg(test)] use ...` block accordingly.

Adjustments:

  * verifier_orchestration.rs: added imports for
    - super::repair_driver: VERIFIER_REPAIR_PASS_MAX_EDITS,
      VERIFIER_REPAIR_PASS_MAX_OUTPUT_BYTES,
      VERIFIER_REPAIR_PASS_MAX_REASON_CHARS (consts the moved
      verifier_repair_intent_limits references)
    - super::repair_patch_validation: VerifierRepairIntentLimits
      (return type of verifier_repair_intent_limits)
    - super::turn::missing_verifier_setup_hint_for_request (turn.rs
      helper called from verifier_setup_policy_message; promoted to
      pub(super) in turn.rs)
    - #[cfg(test)]-gated imports for the dependencies of the 5
      test-only helpers (VerifierRepairDecision, RepairIntentEdit,
      VerifierRepairIntent, repair_intent_edits_fingerprint,
      RecoveryTargetHint, workspace_relative_path_for_tool_arg,
      decision_target_path)

  * turn.rs:
    - Promoted missing_verifier_setup_hint_for_request and
      decision_target_path to pub(super) so verifier_orchestration
      can name them. decision_target_path retains its original
      #[cfg(test)] attribute.
    - Extended use super::verifier_orchestration::{...} with the 3
      production-callable fns; added separate
      #[cfg(test)] use super::verifier_orchestration::{...} for the
      3 test-callable fns (verifier_repair_intent_fingerprint /
      verifier_repair_intents_fingerprint /
      verifier_repair_invalid_can_continue) — verifier_repair_-
      decision_targets_hint and repair_intent_edit_payloads have no
      direct call sites in turn.rs (only the moved fns reference
      them) so they are pub(super) but not re-imported.
    - Removed VERIFIER_REPAIR_PASS_MAX_OUTPUT_BYTES and
      VERIFIER_REPAIR_PASS_MAX_REASON_CHARS from the
      super::repair_driver re-import (now only used by the new
      module).
    - Cleaned 3 orphaned #[cfg(test)] attribute lines left over the
      old fn deletion sites (they would have cfg-gated the next
      production fn out of release builds otherwise).

Metrics: turn.rs 34,833 -> 34,742 (-91); verifier_orchestration.rs
181 -> 296 (+115).
